python/requirements.txt
dlsite-async==0.8.5
requests==2.32.5
pillow==11.3.0